refactor(gateway): extract session API runtime handlers (#2975)
